chore: move safety analysis core to service
