+++ mainappsrc/__init__.py
@@ -40,7 +40,8 @@ _submodule_map = {
     "core.node_utils": "gui.utils.node_utils",
     "fmea_service": "services.safety_analysis.safety_analysis_service",
     "core.fmea_service": "services.safety_analysis.safety_analysis_service",
-    "safety_analysis": "core.safety_analysis",
+    "safety_analysis": "services.safety_analysis.safety_analysis_service",
+    "core.safety_analysis": "services.safety_analysis.safety_analysis_service",
     "automl_core": "core.automl_core",
     # Managers
     "cta_manager": "managers.cta_manager",
